#275 reorganize group tests
