+++ src/bud/test_acct/test_group.py
@@ -1,6 +1,86 @@
+from src._road.finance import default_fund_coin_if_none
+from src._road.road import (
+    get_default_real_id_roadnode as root_label,
+    create_road,
+    default_road_delimiter_if_none,
+)
 from src.bud.group import groupship_shop
-from src.bud.group import groupbox_shop
+from src.bud.group import groupbox_shop, GroupBox
 from pytest import raises as pytest_raises
+
+
+def test_GroupBox_exists():
+    # ESTABLISH
+    swim_text = ",swimmers"
+    # WHEN
+    swim_groupbox = GroupBox(group_id=swim_text)
+    # THEN
+    assert swim_groupbox is not None
+    assert swim_groupbox.group_id == swim_text
+    assert swim_groupbox._groupships is None
+    assert swim_groupbox._fund_give is None
+    assert swim_groupbox._fund_take is None
+    assert swim_groupbox._fund_agenda_give is None
+    assert swim_groupbox._fund_agenda_take is None
+    assert swim_groupbox._credor_pool is None
+    assert swim_groupbox._debtor_pool is None
+    assert swim_groupbox._road_delimiter is None
+    assert swim_groupbox._fund_coin is None
+
+
+def test_groupbox_shop_ReturnsCorrectObj():
+    # ESTABLISH
+    swim_text = ",swimmers"
+    nation_road = create_road(root_label(), "nation-states")
+    usa_road = create_road(nation_road, "USA")
+
+    # WHEN
+    swim_groupbox = groupbox_shop(group_id=swim_text)
+
+    # THEN
+    print(f"{swim_text}")
+    assert swim_groupbox is not None
+    assert swim_groupbox.group_id is not None
+    assert swim_groupbox.group_id == swim_text
+    assert swim_groupbox._groupships == {}
+    assert swim_groupbox._fund_give == 0
+    assert swim_groupbox._fund_take == 0
+    assert swim_groupbox._fund_agenda_give == 0
+    assert swim_groupbox._fund_agenda_take == 0
+    assert swim_groupbox._credor_pool == 0
+    assert swim_groupbox._debtor_pool == 0
+    assert swim_groupbox._road_delimiter == default_road_delimiter_if_none()
+    assert swim_groupbox._fund_coin == default_fund_coin_if_none()
+
+
+def test_groupbox_shop_ReturnsCorrectObj_road_delimiter():
+    # ESTABLISH
+    swim_text = "/swimmers"
+    slash_text = "/"
+    x_fund_coin = 7
+
+    # WHEN
+    swim_groupbox = groupbox_shop(
+        group_id=swim_text, _road_delimiter=slash_text, _fund_coin=x_fund_coin
+    )
+
+    # THEN
+    assert swim_groupbox._road_delimiter == slash_text
+    assert swim_groupbox._fund_coin == x_fund_coin
+
+
+# def test_GroupBox_set_group_id_RaisesErrorIfParameterContains_road_delimiter_And_acct_mirror_True():
+#     # ESTABLISH
+#     slash_text = "/"
+#     bob_text = f"Bob{slash_text}Texas"
+
+#     # WHEN / THEN
+#     with pytest_raises(Exception) as excinfo:
+#         groupbox_shop(bob_text, _acct_mirror=True, _road_delimiter=slash_text)
+#     assert (
+#         str(excinfo.value)
+#         == f"'{bob_text}' needs to be a RoadNode. Cannot contain delimiter: '{slash_text}'"
+#     )
 
 
 def test_GroupBox_set_groupship_CorrectlySetsAttr():
